Added TileView, InteractableView and Level classes

+++ src/Game/Window/Window.h
@@ -14,7 +14,7 @@ class Window {
 
 		void resize(int width, int height);
 
-		static const Uint16 BASE_WINDOW_HEIGHT = 720;
-		static const Uint16 BASE_WINDOW_WIDTH = 1280;
+		static const Uint32 BASE_WINDOW_HEIGHT = 720;
+		static const Uint32 BASE_WINDOW_WIDTH = 1280;
 };
